feat: Refactor TODO app functionality into separate module and enhance utility functions. Side note: Very buggy right now.
