+++ main.py
@@ -10,113 +10,19 @@ import random
 import io
 import socket
 import speedtest
-# Typing-effect print override
-import builtins
-_original_print = builtins.print
-
-def print(*args, sep=' ', end='\n', typing_speed=0, **kwargs):
-    """
-    Typing-effect print: prints each character with a small delay.
-    :param sep: string to separate the arguments
-    :param end: string to append at the end of the print
-    :param typing_speed: delay in seconds between characters
-    """
-    # TODO: Fix the docstring.
-    if typing_speed <= 0:
-        _original_print(*args, sep=sep, end=end, **kwargs)
-        return
-    text = sep.join(str(a) for a in args) + end
-    for ch in text:
-        sys.stdout.write(ch)
-        sys.stdout.flush()
-        time.sleep(typing_speed)
-
 from simpleeval import simple_eval
-
-
-try:
-    import pywhatkit as kit
-except:
-    print('No internet connection')
-
 from gtts import gTTS
 from googletrans import Translator
 import asyncio
 import Levenshtein
 import pyautogui
 
+from todo_app import *
 
-class Colors:
-    """
-    This class contains color codes for terminal text formatting.
-    """
-    RESET = '\033[0m'
-    RED = '\033[31m'
-    GREEN = '\033[32m'
-    YELLOW = '\033[33m'
-    LIGHT_GRAY = '\033[37m'
 
-help_content = ('Type "todo help" to see the commands for the TODO app.'
-              '\nType "check help" to see the commands for the Checklist app.'
-              '\nType open <App Name> to open the desired application.'
-              '\nType "exit" to exit the program.'
-              '\nType "chat" to access the LLM.'
-              '\nType eval <expression> to evaluate a mathematical expression.'
-              '\nType "tts help" to see the commands for the Text-to-Speech app.'
-              '\nType "tr help" to see the commands for the Translation app.'
-              '\nType "settings help" to see the commands for the Settings app.'
-              '\nType "animate" or "animation" or "anim" to see the animated logo.'
-              '\nType "clear" or "clr" to clear the screen.'
-              '\nType "coin" to flip a coin.'
-              '\nType "quit" to exit the program.')
+from utils import *
 
-# Color theme of the program:
-neon_colors = ["\033[35m", "\033[95m","\033[94m",  "\033[94m"]
-yellow_colors = [
-    "\033[93m",           # Bright yellow (standard)
-    "\033[38;5;220m",     # Goldenrod / amber
-    "\033[38;5;214m",     # Deep orange / tangerine
-    "\033[38;5;226m"      # Neon lemon yellow
-]
 
-maintext = r"""
-██████  ██╗╔═██ ╔███    ███╗
-██═╬═██ ██╚╝██╝ ║████  ████║
-██████  █████   ║██╗████╔██║      El Ayuntade By:
-██═╬═██ ██╔╗██╗ ║██╚╗██╔╝██║     Batu Koray Masak
-██████  ██╝╚═██ ╚██ ╚══╝ ██╝
-
-Type "help" to see the available commands.
-
-"""
-goodbye_text = 'Goodbye! | El Ayuntade By: Batu Koray Masak'
-
-def neon_text(text,randomness=True,neon_map_num = 0,colors='neon'):
-    """
-This function takes a text input and returns it with neon colors applied to each character.
-    :param text: The text to be colored.
-    :param randomness: If True, each character will be colored randomly from the neon_colors list. It is true in default.
-    :param neon_map_num: An integer used to map the colors in a specific order. Is 0 in default. Only effective if randomness is False.
-    :param colors: The colors to use. Default is 'neon'. Choices are 'neon' or 'yellow'. TODO: Add more colors.
-    :return: A string with neon colors applied to each character.
-    """
-    if randomness:
-        if colors == 'neon':
-            return ''.join(f"{random.choice(neon_colors)}{char}"for char in text) + Colors.RESET
-        elif colors == 'yellow':
-            return ''.join(f"{random.choice(yellow_colors)}{char}"for char in text) + Colors.RESET
-    else:
-        if colors == 'neon':
-            return ''.join(f"{neon_colors[(j-neon_map_num) % len(neon_colors)]}{char}" for j, char in enumerate(text)) + Colors.RESET
-        elif colors == 'yellow':
-            return ''.join(f"{yellow_colors[(j-neon_map_num) % len(yellow_colors)]}{char}" for j, char in enumerate(text)) + Colors.RESET
-
-commands = ['todo','todo ls','todo add','help','exit','chat','quit','open','todo rm','todo changeorder',
-            'todo abcorder','todo cbaorder','todo do', 'todo help', 'todo add', 'todo ls', 'todo rm all',
-            'eval','clear','clr','open', 'check', 'check ls', 'check add', 'check rm', 'check help',
-            'check -check', 'check -uncheck', 'animate', 'animation', 'anim','you found the easter egg!',
-            'tts','tr','send','settings','settings edit','settings help','settings ls','settings add','settings reset',
-            'notes','notes add','notes rm','notes ls']
 
 def set_command_variables(text_input:str) -> []:
     """
@@ -328,6 +234,7 @@ def clear_screen(text = True,randomness=True,clear_technique='os',subtitle=None)
     :param text: If True, it will print the main text after clearing the screen. If False, it will not print the main text/branding. True in default.
     :param randomness: If True, the main text will be colored randomly. If False, it will use a fixed color pattern. True in default.
     :param clear_technique: The technique to clear the screen. 'os' for using os.system, 'ascii' for using ANSI escape codes. "os" in default.
+    :param subtitle: If True, the main text will be colored subtitle. True in default.
     :return: void
     """
     if clear_technique == 'os':
@@ -350,249 +257,7 @@ def clear_screen(text = True,randomness=True,clear_technique='os',subtitle=None)
         raise Exception(f'The clear technique "{clear_technique}" is not supported.')
 
 
-# TODO App:
-todo_list = []
-
-def update_todo_list():
-    """
-    This function updates the TODO list by reading from a JSON file.
-    :return: void
-    """
-    global todo_list
-    if os.path.exists(user_data.TODO_FILE_LOC):
-        with open(user_data.TODO_FILE_LOC, "r", encoding="utf-8") as f:
-            try:
-                todo_list = json.load(f)
-            except json.JSONDecodeError:
-                todo_list = []
-    else:
-        todo_list = []
-
-update_todo_list()
-
-def todo_save():
-    """
-    This function saves the current TODO list to a file in JSON format.
-    :return: void
-    """
-    with open(user_data.TODO_FILE_LOC, "w", encoding="utf-8") as f:
-        json.dump(todo_list, f, ensure_ascii=False, indent=2)
-
-def todo_help():
-    """
-    This function displays the help content for the TODO app.
-    :return: void
-    """
-
-    print('Type "todo ls" for viewing the TODO list'
-        '\nType "todo add <new TODO element>" to add new todo element to the list.'
-        '\nType "todo rm <Desired Target>" to delete the desired target.'
-        '\nType "todo rm <Desired Index>" to delete the desired index.'
-        '\nType "todo rm all" to delete all items from the TODO list.'
-        '\nType "todo rm" to view deleting todo elements'
-        '\nType "todo changeorder" to change the order of two TODO elements.'
-        '\nType "todo abcorder" to sort the TODO list in alphabetical order.'
-        '\nType "todo cbaorder" to sort the TODO list in reverse alphabetical order.')
-
-def todo_list_view():
-    """
-    This function displays the current content of the TODO list.
-    :return: void
-    """
-    update_todo_list()
-    if not todo_list:
-        print('Your TODO list is empty.')
-        return
-    print('My TODO List Content:')
-    for j in range(len(todo_list)):
-        print(f'{j+1}: {todo_list[j]}')
-
-def todo_add(command_original:str):
-    """
-    This function adds a new item to the TODO list based on the command input.
-    :param command_original: The original command input by the user without multiple whitespaces.
-    :return: void
-    """
-    item = command_original[len('todo add '):]
-    if item.isdigit():
-        print(f'{Colors.RED}Error: The item name cannot be a number.{Colors.RESET}')
-        return
-    if item not in todo_list and item != '':
-        todo_list.append(item)
-        print(f'Added new TODO item: {item}')
-        todo_save()
-    elif command_original[9:] == '':
-        print(f"{Colors.RED}Error: You need to provide a name for the TODO item.{Colors.RESET}")
-    else:
-        print(f'{Colors.RED}Error: The item "{item}" already exists in your TODO list.{Colors.RESET}')
-
-def todo_delete_function(command_original:str):
-    """
-    This function deletes an item from the TODO list based on the command input.
-    :param command_original: The original command input by the user without multiple whitespaces.
-    :return: void
-    """
-    global todo_list
-    command_lower = command_original.lower()
-
-    if not len(todo_list) == 0:
-        if command_lower[len('todo rm '):].isdigit():
-            try:
-                index = int(command_original[8:]) - 1
-                if 0 <= index < len(todo_list):
-                    item = todo_list[index]
-                    todo_list.remove(item)
-                    print(f'Item "{item}" was deleted.')
-                    todo_save()
-                else:
-                    print(f'{Colors.RED}Error: Index out of range.{Colors.RESET}')
-            except ValueError:
-                print(f'{Colors.RED}Error: Invalid input. Please enter a valid index.{Colors.RESET}')
-            return
-        if command_lower != 'todo rm':
-            try:
-                todo_list.remove(command_original[8:])
-            except ValueError:
-
-                try:
-                    todo_list.remove(command_lower[8:])
-                except ValueError:
-                    try:
-                        todo_list.remove(command_original[8:].capitalize())
-                    except ValueError:
-                        print(f'{Colors.RED}{command_lower[8:]} was not found.{Colors.RESET}')
-                    else:
-                        print(f'{command_original[8:]} was deleted.')
-                else:
-                    print(f'{command_original[8:]} was deleted.')
-            else:
-                print(f'{command_original[8:]} was deleted.')
-        else:
-            todo_list_view()
-            indexes = input(  '\nType the index/indexes of the TODO list content or the name of the item that you want to delete. '
-                              '\nSeparate the desired indexes with ",".'
-                              '\nType "all" to delete all items.'
-                              '\nType "exit" if you want to exit.\n')
-
-            if not indexes.lower() == 'exit' and not indexes.lower() == 'all':
-                indexes = indexes.split(',')
-                indexes.sort(reverse=True)
-                try:
-                    if int(max(indexes)) > len(todo_list) or int(min(indexes)) < 1:
-                        print(f'{Colors.RED}Error: The index you are trying to delete is out of range.{Colors.RESET}')
-                        return
-                except ValueError:
-                    print(f'{Colors.RED}Error: Invalid input.{Colors.RESET}')
-                    return
-
-                for index in indexes:
-                    try:
-                        item = todo_list[int(index) - 1]
-                        if 0 <= int(index) <= len(todo_list):
-                            todo_list.remove(todo_list[int(index)-1])
-                    except (IndexError, ValueError):
-                        print(f'{Colors.RED}Error: The item you are trying to delete does not exist{Colors.RESET}')
-                    else:
-                        print(f'Item  {item} was deleted.')
-            elif indexes.lower() == 'all':
-                todo_list.clear()
-                print('All items were deleted.')
-    else:
-        print('Your TODO list is empty.')
-    todo_save()
-
-def todo_changeorder():
-    """
-    This function changes the order of two items in the TODO list based on user input.
-    :return: void
-    """
-    if len(todo_list) < 2:
-        print(f'{Colors.RED}You need at least two items in your TODO list to change their order.{Colors.RESET}')
-        return
-    todo_list_view()
-    userinput = input('Type the indexes of two TODO items you want to swap, \nseparated by a comma(e.g., "1,2"): ')
-
-    num1 = userinput.split(',')[0].strip()
-    num2 = userinput.split(',')[1].strip()
-    try:
-        temp = todo_list[int(num1) - 1]
-        todo_list[int(num1) - 1] = todo_list[int(num2) - 1]
-        todo_list[int(num2) - 1] = temp
-        todo_save()
-        print(f'Swapped items at indexes {num1} and {num2}.')
-    except (IndexError, ValueError):
-        print(f'{Colors.RED}Error: Invalid input. Please enter valid indexes.{Colors.RESET}')
-
-def todo_abcorder():
-    """
-    This function sorts the TODO list in alphabetical order.
-    :return: void
-    """
-    global todo_list
-    todo_list = sorted(todo_list, key=lambda x: x.lower())
-    todo_save()
-    print('TODO list sorted in alphabetical order.')
-
-def todo_cbaorder():
-    """
-    This function sorts the TODO list in reverse alphabetical order.
-    :return: void
-    """
-    global todo_list
-    todo_list = sorted(todo_list, key=lambda x: x.lower(), reverse=True)
-    todo_save()
-    print('TODO list sorted in reverse alphabetical order.')
-
-def todo_do_function():
-    """
-    This function allows the user to work on a TODO item for a specified amount of time.
-    :return: void
-    """
-    todo_list_view()
-
-    random1 = random.randint(1,10)
-    random2 = random.randint(1,90)
-    user_input = input('Type the index of the TODO item you are going to work on, and separate minute time with a comma'
-                   f'\n(Ex: Type "{random1}, {random2}" for doing the activity at index {random1} for {random2} minutes.)'
-                     '\nType "exit" to exit.\n')
-
-    if user_input.lower() == 'exit':
-        return
-    try:
-        user_input = user_input.strip().split(',')
-        index = int(user_input[0]) - 1
-        time_minutes = int(user_input[1])
-        if index < 0 or index >= len(todo_list):
-            print(f'{Colors.RED}Error: Index out of range. Please enter a valid index.{Colors.RESET}')
-            return
-        if time_minutes <= 0:
-            print(f'{Colors.RED}Error: Time must be a positive integer.{Colors.RESET}')
-            return
-
-        for j in range(time_minutes):
-            for k in range(240):
-                print(neon_text(f'Starting work on "{todo_list[index]}" for {time_minutes} minutes.\nMinutes: {j}, Seconds: {int(k/4)}, Percentage: {((j * 60 + k/4) / (time_minutes * 60)) * 100:.2f}%'))
-                time.sleep(0.25)
-                clear_last_lines(2)
-        clear_last_lines(1)
-        print(neon_text(f'Finished working on "{todo_list[index]}".'))
-        write_worklogs(f'{user_data.username} has finished working on the topic {todo_list[index]} for {time_minutes} minutes.')
-        todo_list.pop(index)
-        todo_save()
-
-    except (ValueError, IndexError):
-        print(f'{Colors.RED}Error: Invalid input format. Please use the format "index,time".{Colors.RESET}')
-        return
-
-def write_worklogs(message:str):
-    """
-    This function writes a message to the worklogs file.
-    :param message: The message to be written to the worklogs file.
-    :return: void
-    """
-    with open(user_data.WORKLOGS_FILE_LOC, "a") as f:
-        f.write(message + '\n')
-
+# TODO APP WAS HERE.
 
 # Checklist App:
 checklist_dict = {}
@@ -1107,7 +772,7 @@ def chat_function():
               'Avoid long answers; but if the user asks for a long answer, you can give it.\n'
               'You are an assistant, and you will answer the user\'s questions in a friendly manner.\n')
 
-    print(neon_text('LLM chat mode activated. Type "exit" or "quit" to exit the LLM.'),typing_speed=0.001)
+    print(neon_text('LLM chat mode activated. Type "exit" or "quit" to exit the LLM.'))
     while True:
         try:
             user_input = input(f'{neon_text('You:')} {Colors.LIGHT_GRAY}')
@@ -1124,14 +789,14 @@ def chat_function():
                 subprocess.run(["pkill", "-f", 'Ollama 2'], check=True)
             except subprocess.CalledProcessError:
                 continue
-            print("Exiting chat mode.",typing_speed=0.01)
+            print("Exiting chat mode.")
             break
         try:
-            print(neon_text(random.choice(waiting_messages)),typing_speed=0.001)
+            print(neon_text(random.choice(waiting_messages)))
             ollama_answer = get_ollama_response(f'{prompt} {user_input}')
             chat_logs_llm += 'AI: ' + remove_all_thinks(ollama_answer) + "\n"
             clear_last_lines(1)
-            print(f"{neon_text('AI:')} {Colors.LIGHT_GRAY}{remove_all_thinks(ollama_answer)}{Colors.RESET}",typing_speed=0.004)
+            print(f"{neon_text('AI:')} {Colors.LIGHT_GRAY}{remove_all_thinks(ollama_answer)}{Colors.RESET}")
         except subprocess.CalledProcessError:
             print(f"{Colors.RED}Error: Chat command failed.{Colors.RESET}")
 
@@ -1207,7 +872,7 @@ def translate_function(command_original: str):
                 return await tr.translate(t, dest=lang)
             result = asyncio.run(_do_translate(text, 'en')).text
             print(f'En: "{neon_text(result)}"')
-        except Exception as e:
+        except Exception:
             print(f'{Colors.RED}Internet error: Possible lossy internet connection{Colors.RESET}')
         return
 
@@ -1276,7 +941,7 @@ def is_connected_socket(host="8.8.8.8", port=53, timeout=3):
 
 
 def measure_speed():
-    print(neon_text('Measuring internet speed...(25 seconds)'),typing_speed=0.00025)
+    print(neon_text('Measuring internet speed...(25 seconds)'))
     successful = True
     try:
         st = speedtest.Speedtest()
@@ -1286,6 +951,7 @@ def measure_speed():
         ping      = st.results.ping        # ms
     except KeyboardInterrupt:
         print(f'{Colors.RED}Internet connection interrupted.{Colors.RESET}')
+        successful = False
     except Exception as e:
         print(f'{Colors.RED}Internet connection error. {Colors.RESET}')
         successful = False
@@ -1309,7 +975,6 @@ def unknown_command(command_original, app_name=None):
     This function handles unknown commands by suggesting the closest command
     from the predefined list, using Levenshtein distance.
     """
-    typing_speed_default = 0.005
     parts = command_original.strip().split()
     if not parts:
         return
@@ -1317,28 +982,28 @@ def unknown_command(command_original, app_name=None):
     if app_name is None:
         # Compare full commands
         closest = min(commands, key=lambda cmd: levenshtein(command_original, cmd))
-        print(f'{Colors.RED}Unknown command: "{parts[0]}". Did you mean "{closest}"?{Colors.RESET}',typing_speed=typing_speed_default)
+        print(f'{Colors.RED}Unknown command: "{parts[0]}". Did you mean "{closest}"?{Colors.RESET}')
 
     elif app_name == 'todo':
         # Suggest a todo subcommand
         pool = [c.split(' ', 1)[1] for c in commands if c.startswith('todo ')]
         target = parts[1] if len(parts) > 1 else ''
         closest = min(pool, key=lambda sub: levenshtein(target, sub))
-        print(f'{Colors.RED}Unknown TODO command: "{target}". Did you mean "todo {closest}"?{Colors.RESET}', typing_speed=typing_speed_default)
+        print(f'{Colors.RED}Unknown TODO command: "{target}". Did you mean "todo {closest}"?{Colors.RESET}')
 
     elif app_name == 'check':
         # Suggest a checklist subcommand
         pool = [c.split(' ', 1)[1] for c in commands if c.startswith('check ')]
         target = parts[1] if len(parts) > 1 else ''
         closest = min(pool, key=lambda sub: levenshtein(target, sub))
-        print(f'{Colors.RED}Unknown checklist command: "{target}". Did you mean "check {closest}"?{Colors.RESET}', typing_speed=typing_speed_default)
+        print(f'{Colors.RED}Unknown checklist command: "{target}". Did you mean "check {closest}"?{Colors.RESET}')
 
     elif app_name == 'settings':
         # Suggest a settings subcommand
         pool = [c.split(' ', 1)[1] for c in commands if c.startswith('settings ')]
         target = parts[1] if len(parts) > 1 else ''
         closest = min(pool, key=lambda sub: levenshtein(target, sub))
-        print(f'{Colors.RED}Unknown settings command: "{target}". Did you mean "settings {closest}"?{Colors.RESET}', typing_speed=typing_speed_default)
+        print(f'{Colors.RED}Unknown settings command: "{target}". Did you mean "settings {closest}"?{Colors.RESET}')
 
 def main():
     """
